feat(API): Queue tasks via host-server and execute tasks scheduled for the current node when queued on the server
This allows the user to queue any number of task for any number of
projects, and this server will create command references for the target
node.
If a command is registered for this node, then this node will execute
that command, and execute any related commands (in order) as each
command completes succesfully.